feat(sustainability): real envelope conduction + per-façade solar (WS A2)
The annual energy estimate now runs off a REAL envelope instead of
internal-gains only — closing the "no envelope data, gains under-counted"
gap.

- Extract EnvelopeDetector (Core/Hvac/Loads) from HvacBlockLoadCommand:
  boundary→exterior-wall+hosted-window net area + orientation + top-level
  roof, with the generic-ratio fallback. One detector, no fork.
- HvacBlockLoadCommand now delegates to it (behaviour identical); its
  InvalidateTopLevelCache forwards to the shared cache so the document-close
  hook is unchanged.
- SustainabilityEngine resolves the project ConstructionProfile (tunable
  U-values/SHGC) and populates each Space/Room zone's envelope via the shared
  detector, so conduction + per-façade orientation solar are counted; the
  estimator still flags any zone with no derivable envelope.
- Tests: +3 (temperate vs tropical EUI differ; envelope U-value drives
  conduction load; glazing SHGC drives solar cooling). 133 → 136 green.
